+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Upcoming 7.0.0.alpha
 
 - **Breaking change**: Require Ruby >= 3.1 and I18n ~> 1.9
+- **Breaking change**: Remove deprecated methods:
+  - `Money.infinite_precision`
+  - `Money.infinite_precision=`
 - **Potential breaking change**: Fix RSD (Serbian Dinar) formatting to be like `12.345,42 RSD`
 - **Potential breaking change**: Fix USDC decimals places from 2 to 6
 - **Potential breaking change**: Fix MGA (Malagasy Ariary) to be a zero-decimal currency (changing subunit_to_unit from 5 to 1)